fix: Raster Circle render when existing components are queued
This fixes a regression in v0.5.6 when the Circle render was fixed to be more circular and less square.  A hidden behavior in ArcTo would result in a `LineTo` operation rather than `MoveTo`.
Although this behavior is somewhat suspect, I am doing a more minimal fix here to just do a proactive `MoveTo` operation, and optimizing `LineTo` and `MoveTo` no-ops.
Fixes #78